Major changes to code organisation
