+++ kinetics/__init__.py
@@ -1,7 +1,13 @@
-from kinetics.Model import Model, Metrics, uM_to_mgml
-from kinetics.Genetic_Algorithm import GA_Base_Class
-from kinetics.Reaction_Classes import *
-from kinetics import Uncertainty, Sampling
+from kinetics.Model import Model
+from kinetics.reaction_classes import *
+
+from kinetics.optimisation.metrics import Metrics, uM_to_mgml
+from kinetics.optimisation.genetic_algorithm import GA_Base_Class
+
+from kinetics.ua_and_sa.sampling import sample_distributions, sample_uniforms, salib_problem_with_bounds, make_saltelli_samples
+from kinetics.ua_and_sa.run_all_models import run_all_models, dataframes_all_runs, dataframes_quartiles
+from kinetics.ua_and_sa.plotting import plot_substrate, plot_ci_intervals, plot_data, remove_st_less_than, plot_sa_total_sensitivity
+from kinetics.ua_and_sa.sensitivity_analysis import get_concentrations_at_timepoint, get_time_to_concentration, analyse_sobal_sensitivity
 
 
-__version__ = '1.3.6'
+__version__ = '1.3.7'
