+++ kinetics/other_analysis/initial_rates.py
